+++ src/main/kotlin/astminer/config/PipelineConfig.kt
@@ -16,7 +16,7 @@ data class PipelineConfig(
     val filters: List<FilterConfig> = emptyList(),
     @SerialName("label") val labelExtractor: LabelExtractorConfig,
     val storage: StorageConfig,
-    val metadata: MetaDataConfig,
+    val metadata: MetaDataConfig = MetaDataConfig(),
     val numOfThreads: Int = 1,
     val compressBeforeSaving: Boolean = false
 ) {
