+++ src/main/kotlin/astminer/storage/path/PathBasedStorage.kt
@@ -6,11 +6,6 @@ import astminer.common.model.Storage
 import astminer.paths.PathMiner
 import astminer.paths.PathRetrievalSettings
 import astminer.paths.toPathContext
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.buildJsonObject
-import kotlinx.serialization.json.encodeToJsonElement
-import kotlinx.serialization.json.put
 import java.io.File
 import java.io.PrintWriter
 import kotlin.io.path.Path
